+++ src/components/budgetTracker/BudgetTrackerKernel.java
@@ -23,62 +23,14 @@ public interface BudgetTrackerKernel
     float ROUNDING = 100.0f;
 
     /**
-     * Inner class representing a single budget record. It stores the amount of
-     * money and its description.
+     * A record representing a transaction with an amount and a description.
+     *
+     * @param amount
+     *            the amount of money of the record
+     * @param description
+     *            the description of the record
      */
-    class Record {
-        /*
-         * Private members
-         */
-
-        /**
-         * The amount of money.
-         */
-        private float amount;
-
-        /**
-         * The record description.
-         */
-        private String description;
-
-        /*
-         * Constructors
-         */
-
-        /**
-         * Constructor for a record.
-         *
-         * @param amount
-         *            the amount of money
-         * @param description
-         *            the description
-         */
-        public Record(float amount, String description) {
-            this.amount = amount;
-            this.description = description;
-        }
-
-        /*
-         * Public methods
-         */
-
-        /**
-         * Returns the amount of the record.
-         *
-         * @return the amount of the record
-         */
-        public float amount() {
-            return this.amount;
-        }
-
-        /**
-         * Returns the description of the record.
-         *
-         * @return the description of the record
-         */
-        public String description() {
-            return this.description;
-        }
+    record Transaction(float amount, String description) {
     }
 
     /**
@@ -97,7 +49,7 @@ public interface BudgetTrackerKernel
      * @requires |this| > 0
      * @ensures #this = this * <remove>
      */
-    Record remove();
+    Transaction remove();
 
     /**
      * Add a record to the tracker.
@@ -119,7 +71,7 @@ public interface BudgetTrackerKernel
      * @update this
      * @ensures this = #this * <record>
      */
-    void add(Record record);
+    void add(Transaction record);
 
     /**
      * Reports the current balance, rounded according to {@code ROUNDING}.
@@ -128,4 +80,22 @@ public interface BudgetTrackerKernel
      * @ensures balance = [sum of all amounts in the tracker]
      */
     float balance();
+
+    /**
+     * Sets the name of the account.
+     *
+     * @param name
+     *            the name of the account
+     * @update this
+     * @ensures this.accountName = name
+     */
+    void setAccountName(String name);
+
+    /**
+     * Returns the name of the account.
+     *
+     * @return the name of the account
+     * @ensures getAccountName = this.accountName
+     */
+    String getAccountName();
 }